make viewmodel api consist across platform
