+++ precompose-viewmodel/src/commonMain/kotlin/moe/tlaster/precompose/viewmodel/ViewModelAdapter.kt
@@ -51,3 +51,15 @@ private fun <T : ViewModel> StateHolder.getViewModel(
         creator()
     }
 }
+
+/**
+ * Returns a [ViewModel] instance that is scoped to the given [StateHolder].
+ * @param keys A list of keys that will be used to identify the ViewModel.
+ * @param creator A function that will be used to create the ViewModel if it doesn't exist.
+ * @return A ViewModel instance.
+ */
+@Composable
+inline fun <reified T : ViewModel> viewModel(
+    keys: List<Any?> = emptyList(),
+    noinline creator: (SavedStateHolder) -> T,
+): T = viewModel(T::class, keys, creator = creator)
